Les commentaires et les posts ont la PDP de chaque USER qui s'affiche

+++ front-end/src/components/EditProfilModal/EditProfilModal.js
@@ -43,6 +43,8 @@ const EditProfilModal = () => {
   const history = useHistory();
 
   const desactivateAccount = () => {
+    if (!window.confirm(`Voulez-vous vraiment désactiver le compte ?`)) return;
+
     const userId = JSON.parse(localStorage.getItem("user")).user_id;
     axios.get(`http://localhost:4200/api/auth/desactivateAccount/${userId}`);
     localStorage.clear();
